requirements.txt
easyocr==1.7.2
#Pillow==11.1.0
#python-decouple==3.8

Requests==2.32.3
supabase==2.11.0

psycopg2-binary==2.9.10
python-dotenv
sqlalchemy
alembic

pre-commit